changed all index-counters to size_t for improve work os x86_64 systems

+++ includes/output/default_format/ls_df.h
@@ -6,7 +6,7 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/22 12:49:55 by tmaluh            #+#    #+#             */
-/*   Updated: 2020/01/03 20:21:56 by tmaluh           ###   ########.fr       */
+/*   Updated: 2020/01/26 22:53:47 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,6 @@
 /*
 ** PDF - Print Default Format
 */
-extern void
-pdf(const int32_t n_objs, const struct s_object *const objs);
+extern void	pdf(const size_t n_objs, const struct s_object *const objs);
 
 #endif
