+++ includes/parsers/ls_parsers.h
@@ -6,7 +6,7 @@
 /*   By: tmaluh <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/22 13:03:17 by tmaluh            #+#    #+#             */
-/*   Updated: 2020/01/03 20:21:25 by tmaluh           ###   ########.fr       */
+/*   Updated: 2020/01/26 22:52:52 by tmaluh           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ extern int
 parse_args(struct s_arg_av *restrict const a);
 
 extern void
-parse_args_files_as_dir(struct s_arg_obj *args, int32_t n);
+parse_args_files_as_dir(struct s_arg_obj *args, size_t n);
 
 extern int
 parse_flags(struct s_arg_av *restrict const a);
